moved APP.routes('/') from mslib/mswms/wms to docs Blueprint
